apps/password_manager/pyproject.toml
[project]
name = "password-manager"
version = "0.1.0"
requires-python = ">=3.12"
description = "CLI password manager (MVP)"
readme = "README.md"
dependencies = [
  "typer>=0.9.0",
  "cryptography>=41.0",
  "argon2-cffi>=21.3.0",
]

[build-system]
requires = ["setuptools","wheel"]
build-backend = "setuptools.build_meta"
